Parse headers from URL for request
Not tested, and will break tests because it changes signature

Fix header assignment

Tested locally and it works! When using curl you need to encode the cookie becaues of the ;

Now time to fix the tests and get this deployed!

Possibly fix tests

The make file doesn't run them, so I don't know if the pass.
I think it should!

Fallback to env vars for S3

Will only get creds, not location

Fix S3 fallback to not put secret in region

Wow

Update Maintainer to Label